Remove (undocumented) automated signing mode
This mode never quite worked properly due to the signatures being deemed
invalid by Nix; I speculate that this might be due to the CryptEd25519
package from Perl subtly using a different signature format (there's a
bit of history mentioned in issue #1), but haven't confirmed this.
Either way, remove it for now. We can revive it later if needed.

Closes #1.

Signed-off-by: Austin Seipp <aseipp@pobox.com>