eris: handle nar.{xz,bz2} routes as well
The narinfo handler does not advertise .nar.xz or .nar.bz2 files itself,
so normally clients will never hit these routes. However, a future
"upstream" configuration mode will allow Eris to proxy missing requests
to some other cache server, and that server *will* probably serve
narinfos that ask you to fetch a .xz or .bz2 URL. In which case, we need
to handle them appropriately in order to proxy the right request.

Beyond that, supporting these routes is easy enough on its own. In order
to reduce the load as much as possible for concurrent nar requests, we
use the --fast modes for both bzip2 and xz.

Signed-off-by: Austin Seipp <aseipp@pobox.com>