eris: add 'upstream' proxy mode
A new 'upstream' configuration parameter allows you to configure an HTTP
server to be queried on a local cache miss, or in all cases (effectively
a reverse proxy for Nix binary caches).

Currently, only HTTP is supported (perhaps one day, S3 will be supported
as well.)

This mode is EXPERIMENTAL and possibly unstable.

Signed-off-by: Austin Seipp <aseipp@pobox.com>