eris: always use configured signing key hostname
Signing keys for Nix caches are written to files, using the form
"keyname:ed25519 pubkey", where the keyname is (canonically) the URL
pointing to the cache, and the public key is self explanatory.

The key name is ultimately arbitrary and is not used in the
Ed25519/signing calculation in any way, it is just a way to assign an
actual mnemontic to the base64 encoded public key.

For Eris, we allow the user to configure the key name inside the
configuration file, so it can be served to users. However, we did not
account for the fact that the actual keyfiles may have a different name
attached, in fact, $sign_host was mostly unused.

This fixes that discrepancy: if a key name exists and is attached to the
public key, it is stripped, and the configured key name is used instead.

Signed-off-by: Austin Seipp <aseipp@pobox.com>