eris: add 'resign' feature for upstreams
Upstream URLs configured with the '?resign=1' parameter will, if a
signing key is set, fetch narinfos from the upstream server and append
their own computed signature to the result.

The intent is that users can trust a single (possibly private) signing
key used by Eris in proxy/upstream mode, so it can provide a single
endpoint of trust, even for results served by an upstream cache.

Resigning, by default, requires a public key to be configured for the
upstream server, and expects ALL narinfos fetched from the upstream to
be signed with that key. If a signature does not exist, or verification
of the signature wrt the public key fails, then a 404 is returned
immediately.

Signed-off-by: Austin Seipp <aseipp@pobox.com>