eris: automatically compute signing public key
The 'public' setting specifying the public signing key in the
configuration file is redundant: Nix uses ed25519 signing keys for
binary caches, which contain the public key within them. (From a 64-byte
string, the first 32 bytes are the secret key, and the last 32 bytes are
the public key.)

Compute the public key from the private key directly; this makes signing
configurations even easier and simpler. It also makes the logic for
signing setup a bit nicer while we're here.

Signed-off-by: Austin Seipp <aseipp@pobox.com>