eris: fix streaming of upstream .nar files
The ->proxy->start_p() method doesn't seem to be highly reliable; in its
stead, replace it with a simple streaming implementation that just
writes every incoming chunk directly to the client.

This seems to have two deficiencies:

  1) The request is buffered in memory before returning, it seems.

  2) As a result, the headers are not sent until the full request is
buffered, whereas we should stream the headers nearly instantly once we
know the Content-Length, and then stream the body

Ideally we should have a non-blocking, constant-memory stream from the
upstream source to the client. Once that is fixed, the upstream mode
should be stable to use.

Signed-off-by: Austin Seipp <aseipp@pobox.com>