eris: ping upstream /nix-cache-info on startup
When an upstream cache is configured, do a blocking GET request to
/nix-cache-info on startup, so we can ensure the StoreDir setting is the
same as the objects we'll serve locally. If it's not, then disable it
and put up an error message.

Technically, if upstream narinfo signatures are validated, then we can
be sure the store directory was the same already: the store dir will be
part of the 'References:' in the calculated fingerprint, so a valid
fingerprint would need to have the same $Nix::Config::storeDir locally.
That said, if this happened, it would have resulted in a bizarre signing
failure, so checking it immediately is smarter.

In theory, this check isn't needed in always-proxy mode, since we do not
actually care what the local storeDir is configured to be. We don't
support this, and doing it would need a few more tweaks (e.g. serving
the upstream 'StoreDir:' in our own /nix-cache-info), but it might be
useful in to add in the future.

Signed-off-by: Austin Seipp <aseipp@pobox.com>