nix: improve the default.nix configuration
By default, the nix build now uses the nixpkgs revision specified in
nix/nixpkgs.json, whereas before it used <nixpkgs>. So we change that.

Moreover, we want to be able to build against arbitrary versions
nixpkgs, given by HTTP URLs, so we need to distinguish the values given
to us through the 'nixpkgs' parameter and use import/fetchTarball as
needed.

This is also an attempt to unify the logic and make it align with
release.nix. This does not change release.nix, and it ALWAYS uses the
fixed package set without question. That may change in the future too,
however, depending on how we want to publish docker images.

This will (hopefully soon) be used in CI to ensure Eris can be tested
against its stable version, as well as nixpkgs channel revisions.

Signed-off-by: Austin Seipp <aseipp@pobox.com>