eris: experimental json rendering for narinfos
This allows you to request any .narinfo from the server, and, if the
query string '?json=1' is appended at the end, the .narinfo file will be
returned as a JSON object, rather than an ordinary .narinfo manifest.

The intent of this feature is to make it easy to 'scrape' information
about cache objects from other 3rd party tools, without re-creating the
.narinfo parser yourself. For example:

    $ curl -s \
      'http://.../q01x3pcclj1kn7h8v20azbwn1pi4khhm.narinfo?json=1' \
      | jq '.'
    {
      "Compression": "xz",
      "Deriver": "gknag9wnazq1zn4qsg1a2cnjj1a5bfkn-jq-1.6.drv",
      "FileHash": "sha256:1xcwcxrmrras5jpq9x4ysjwrcpdv8c8zmpa40zcf7m6njwm0byjl",
      "FileSize": 9580,
      "NarHash": "sha256:0zffhmwgb97dmw7f3b7wdlmf4w8afc4w7d1qrba58d51mz5bl5qs",
      "NarSize": 36112,
      "References": [
        "/nix/store/6yaj6n8l925xxfbcd65gzqx3dz7idrnn-glibc-2.27",
        "/nix/store/k4hn4wa83wx5blrkyz65s08hi2arfawp-jq-1.6-lib",
        "/nix/store/rxh947ggk80hafg64bc6j2vy6v2l2a78-onig-6.9.3"
      ],
      "Sig": [
        "cache.nixos.org-1:83hlxUyBgP/Aif5nZB4OUmHCoQ5C5QLk4H1IvrS0pxZRXqH3D/pTc96hysJV1SEa2k/D3qnZ4PJa8uPvVqPhAg==",
      ],
      "StorePath": "/nix/store/q01x3pcclj1kn7h8v20azbwn1pi4khhm-jq-1.6-bin",
      "URL": "nar/1xcwcxrmrras5jpq9x4ysjwrcpdv8c8zmpa40zcf7m6njwm0byjl.nar.xz"
    }

(Note: this narinfo, `q01x3pcclj1kn7h8v20azbwn1pi4khhm`, is available in
the upstream cache, so you can verify the above metadata matches.)

Currently this feature is not gated behind any configuration option(s);
it is always enabled in local, proxy, or local+proxy mode.

Signed-off-by: Austin Seipp <aseipp@pobox.com>