nix: use bootstrap.nix for simpler setup
This uses bootstrap.nix to set everything up:

    https://github.com/thoughtpolice/bootstrap.nix

It also has a minor tweak to handle the usage of the lockfile when
passing the argument '--arg nixpkgs channel:lockfile' to the bootstrap
routine. This is to make the CI pipeline look a bit nicer.

It also propagates all the needed options to release.nix to set the
channel properly, where it didn't before.

Signed-off-by: Austin Seipp <aseipp@pobox.com>