eris: EXPERIMENTAL support for Server Push headers
Currently, I'm exploring if Server Push is a viable mechanism for
downloading objects from the Nix cache, using Fastly VCL.

To do this properly, we need to configure a header that contains the URL
we want to push to the user so the VCL can pull it out and push to the
client on demand. (VCL cannot do body processing, at the moment.)

However, doing this in the generalized upstream setting where we upload
to S3 is harder -- we need to tag the object with metadata on upload,
which has to be done in the Nix codebase, and then pull the metadata out
from the header.

Having Eris support this header instead and serve it in all code paths
gives a viable means forward for testing server-side push support in
several settings, befor needing to do Nix integration. In the event Push
support works out, it will be a very useful feature in its own right
behind the right HTTP/2 proxy.

Signed-off-by: Austin Seipp <aseipp@pobox.com>