use write instead of write_chunked
Setting Content-Length to a value, while also setting Transfer-Encoding
to chunked is not valid HTTP. To quote RFC 7230, Section:
"A sender MUST NOT send a Content-Length header field in any message that contains a Transfer-Encoding header field."

While libcurl, nginx and haproxy all seem fine with this, varnish is not. As in, varnish refuses responses from eris.

Signed-off-by: ajs124 <git@ajs124.de>